+++ HappyTesla/Controllers/ReservationsController.cs
@@ -88,8 +88,13 @@ public class ReservationsController : Controller
     [Route("{action}/{reservationId:int}")]
     public async Task<ActionResult> DeleteReservation(int reservationId)
     {
-        await _reservationsRepository.DeleteReservationAsync(reservationId);
-
+        var reservation = await _reservationsRepository.GetReservationByIdAsync(reservationId);
+        
+        if (reservation.StartDate > DateTime.Now)
+        {
+            await _reservationsRepository.DeleteReservationAsync(reservationId);
+        }
+        
         return RedirectToAction("GetAllReservationsForCurrentlyLoggedInUser");
     }
